# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OrderItem {
  name: string;
  quantity: number;
  price: number;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  ready: "ready",
  picked_up: "picked_up",
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  platform: string;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  estimatedPrepTime?: number | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryPersonName?: string | null;
  /** @nullable */
  deliveryPersonPhone?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  acceptedAt?: string | null;
  /** @nullable */
  readyAt?: string | null;
}

export type ListOrdersResponse = Order[];

export interface CreateOrderBody {
  orderNumber: string;
  platform: string;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  estimatedPrepTime?: number | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryPersonName?: string | null;
  /** @nullable */
  deliveryPersonPhone?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface AcceptOrderBody {
  /** @nullable */
  estimatedPrepTime?: number | null;
}

export interface RejectOrderBody {
  /** @nullable */
  reason?: string | null;
}

export interface OrderStats {
  totalToday: number;
  pendingCount: number;
  acceptedCount: number;
  readyCount: number;
  totalRevenue: number;
  /** @nullable */
  avgPrepTime?: number | null;
}

export type ListOrdersParams = {
  status?: string;
  limit?: number;
};