# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all orders
 */
export const ListOrdersQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
});

export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  status: zod.enum(["pending", "accepted", "rejected", "ready", "picked_up"]),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
  rejectionReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  acceptedAt: zod.coerce.date().nullish(),
  readyAt: zod.coerce.date().nullish(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  orderNumber: zod.string(),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a single order
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  status: zod.enum(["pending", "accepted", "rejected", "ready", "picked_up"]),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
  rejectionReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  acceptedAt: zod.coerce.date().nullish(),
  readyAt: zod.coerce.date().nullish(),
});

/**
 * @summary Accept an order (restaurant validates it)
 */
export const AcceptOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const AcceptOrderBody = zod.object({
  estimatedPrepTime: zod.number().nullish(),
});

export const AcceptOrderResponse = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  status: zod.enum(["pending", "accepted", "rejected", "ready", "picked_up"]),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
  rejectionReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  acceptedAt: zod.coerce.date().nullish(),
  readyAt: zod.coerce.date().nullish(),
});

/**
 * @summary Reject an order
 */
export const RejectOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const RejectOrderBody = zod.object({
  reason: zod.string().nullish(),
});

export const RejectOrderResponse = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  status: zod.enum(["pending", "accepted", "rejected", "ready", "picked_up"]),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
  rejectionReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  acceptedAt: zod.coerce.date().nullish(),
  readyAt: zod.coerce.date().nullish(),
});

/**
 * @summary Mark order as ready for pickup
 */
export const MarkOrderReadyParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkOrderReadyResponse = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  status: zod.enum(["pending", "accepted", "rejected", "ready", "picked_up"]),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
  rejectionReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  acceptedAt: zod.coerce.date().nullish(),
  readyAt: zod.coerce.date().nullish(),
});

/**
 * @summary Get order statistics summary
 */
export const GetOrderStatsResponse = zod.object({
  totalToday: zod.number(),
  pendingCount: zod.number(),
  acceptedCount: zod.number(),
  readyCount: zod.number(),
  totalRevenue: zod.number(),
  avgPrepTime: zod.number().nullish(),
});

/**
 * @summary Get the most recent orders
 */
export const GetRecentOrdersResponseItem = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  status: zod.enum(["pending", "accepted", "rejected", "ready", "picked_up"]),
  platform: zod.string(),
  customerName: zod.string(),
  customerPhone: zod.string().nullish(),
  items: zod.array(
    zod.object({
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      notes: zod.string().nullish(),
    }),
  ),
  totalAmount: zod.number(),
  estimatedPrepTime: zod.number().nullish(),
  deliveryAddress: zod.string().nullish(),
  deliveryPersonName: zod.string().nullish(),
  deliveryPersonPhone: zod.string().nullish(),
  notes: zod.string().nullish(),
  rejectionReason: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  acceptedAt: zod.coerce.date().nullish(),
  readyAt: zod.coerce.date().nullish(),
});
export const GetRecentOrdersResponse = zod.array(GetRecentOrdersResponseItem);