# lib/api-zod/src/generated/types/acceptOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */

export interface AcceptOrderBody {
  /** @nullable */
  estimatedPrepTime?: number | null;
}