# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface CreateOrderBody {
  orderNumber: string;
  platform: string;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  estimatedPrepTime?: number | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryPersonName?: string | null;
  /** @nullable */
  deliveryPersonPhone?: string | null;
  /** @nullable */
  notes?: string | null;
}