# lib/api-zod/src/generated/types/listOrdersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

export type ListOrdersResponse = Order[];