# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  platform: string;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  estimatedPrepTime?: number | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryPersonName?: string | null;
  /** @nullable */
  deliveryPersonPhone?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  acceptedAt?: Date | null;
  /** @nullable */
  readyAt?: Date | null;
}