# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  name: string;
  quantity: number;
  price: number;
  /** @nullable */
  notes?: string | null;
}