# lib/api-zod/src/generated/types/orderStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderStats {
  totalToday: number;
  pendingCount: number;
  acceptedCount: number;
  readyCount: number;
  totalRevenue: number;
  /** @nullable */
  avgPrepTime?: number | null;
}