# lib/api-zod/src/generated/types/rejectOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Restaurant Order API
 * OpenAPI spec version: 0.1.0
 */

export interface RejectOrderBody {
  /** @nullable */
  reason?: string | null;
}